topdown: Fix data race in the http.send timeout tests
Also, since these tests are slow, move them into another flag and gate
them with a build flag.
