+++ topdown/http_test.go
@@ -6,7 +6,6 @@ package topdown
 
 import (
 	"bytes"
-	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/json"
@@ -18,7 +17,6 @@ import (
 	"os"
 	"strconv"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
@@ -501,110 +499,6 @@ func TestInvalidKeyError(t *testing.T) {
 	}
 }
 
-func TestHTTPSendTimeout(t *testing.T) {
-
-	// Each test can tweak the response delay, default is 0 with no delay
-	var responseDelay time.Duration
-
-	tsMtx := sync.Mutex{}
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		tsMtx.Lock()
-		defer tsMtx.Unlock()
-		time.Sleep(responseDelay)
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte(`hello`))
-	}))
-	// Note: We don't Close() the test server as it will block waiting for the
-	// timed out clients connections to shut down gracefully (they wont).
-	// We don't need to clean it up nicely for the unit test.
-
-	tests := []struct {
-		note           string
-		rule           string
-		input          string
-		defaultTimeout time.Duration
-		evalTimeout    time.Duration
-		serverDelay    time.Duration
-		expected       interface{}
-	}{
-		{
-			note:     "no timeout",
-			rule:     `p = x { http.send({"method": "get", "url": "%URL%" }, resp); x := remove_headers(resp) }`,
-			expected: `{"body": null, "raw_body": "hello", "status": "200 OK", "status_code": 200}`,
-		},
-		{
-			note:           "default timeout",
-			rule:           `p = x { http.send({"method": "get", "url": "%URL%" }, x) }`,
-			evalTimeout:    1 * time.Minute,
-			serverDelay:    5 * time.Second,
-			defaultTimeout: 500 * time.Millisecond,
-			expected:       &Error{Code: BuiltinErr, Message: "request timed out"},
-		},
-		{
-			note:           "eval timeout",
-			rule:           `p = x { http.send({"method": "get", "url": "%URL%" }, x) }`,
-			evalTimeout:    500 * time.Millisecond,
-			serverDelay:    5 * time.Second,
-			defaultTimeout: 1 * time.Minute,
-			expected:       &Error{Code: BuiltinErr, Message: "context deadline exceeded"},
-		},
-		{
-			note:           "param timeout less than default",
-			rule:           `p = x { http.send({"method": "get", "url": "%URL%", "timeout": "500ms"}, x) }`,
-			evalTimeout:    1 * time.Minute,
-			serverDelay:    5 * time.Second,
-			defaultTimeout: 1 * time.Minute,
-			expected:       &Error{Code: BuiltinErr, Message: "request timed out"},
-		},
-		{
-			note:           "param timeout greater than default",
-			rule:           `p = x { http.send({"method": "get", "url": "%URL%", "timeout": "500ms"}, x) }`,
-			evalTimeout:    1 * time.Minute,
-			serverDelay:    5 * time.Second,
-			defaultTimeout: 1 * time.Millisecond,
-			expected:       &Error{Code: BuiltinErr, Message: "request timed out"},
-		},
-		{
-			note:           "eval timeout less than param",
-			rule:           `p = x { http.send({"method": "get", "url": "%URL%", "timeout": "1m" }, x) }`,
-			evalTimeout:    500 * time.Millisecond,
-			serverDelay:    5 * time.Second,
-			defaultTimeout: 1 * time.Minute,
-			expected:       &Error{Code: BuiltinErr, Message: "context deadline exceeded"},
-		},
-	}
-
-	for _, tc := range tests {
-		responseDelay = tc.serverDelay
-
-		ctx := context.Background()
-		var cancel context.CancelFunc
-		if tc.evalTimeout > 0 {
-			ctx, cancel = context.WithTimeout(ctx, tc.evalTimeout)
-		}
-
-		// TODO(patrick-east): Remove this along with the environment variable so that the "default" can't change
-		originalDefaultTimeout := defaultHTTPRequestTimeout
-		if tc.defaultTimeout > 0 {
-			defaultHTTPRequestTimeout = tc.defaultTimeout
-		}
-
-		rule := strings.ReplaceAll(tc.rule, "%URL%", ts.URL)
-		if e, ok := tc.expected.(*Error); ok {
-			e.Message = strings.ReplaceAll(e.Message, "%URL%", ts.URL)
-		}
-
-		runTopDownTestCaseWithContext(ctx, t, map[string]interface{}{}, tc.note, append(httpSendHelperRules, rule), nil, tc.input, tc.expected)
-
-		// Put back the default (may not have changed)
-		defaultHTTPRequestTimeout = originalDefaultTimeout
-		if cancel != nil {
-			cancel()
-		}
-	}
-}
-
 func TestParseTimeout(t *testing.T) {
 	tests := []struct {
 		note     string
@@ -798,7 +692,7 @@ func TestHTTPSendCaching(t *testing.T) {
 		},
 		{
 			note: "http.send GET cache hit",
-			ruleTemplate: `p = x { 
+			ruleTemplate: `p = x {
 									r1 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})
 									r2 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})  # cached
 									r3 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})  # cached
@@ -811,7 +705,7 @@ func TestHTTPSendCaching(t *testing.T) {
 		},
 		{
 			note: "http.send GET cache miss different method",
-			ruleTemplate: `p = x { 
+			ruleTemplate: `p = x {
 									r1 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})
 									r2 = http.send({"method": "post", "url": "%URL%", "force_json_decode": true})
 									r1_2 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})  # cached
@@ -823,7 +717,7 @@ func TestHTTPSendCaching(t *testing.T) {
 		},
 		{
 			note: "http.send GET cache miss different url",
-			ruleTemplate: `p = x { 
+			ruleTemplate: `p = x {
 									r1 = http.send({"method": "get", "url": "%URL%/foo", "force_json_decode": true})
 									r2 = http.send({"method": "get", "url": "%URL%/bar", "force_json_decode": true})
 									r1_2 = http.send({"method": "get", "url": "%URL%/foo", "force_json_decode": true})  # cached
@@ -835,7 +729,7 @@ func TestHTTPSendCaching(t *testing.T) {
 		},
 		{
 			note: "http.send GET cache miss different decode opt",
-			ruleTemplate: `p = x { 
+			ruleTemplate: `p = x {
 									r1 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})
 									r2 = http.send({"method": "get", "url": "%URL%", "force_json_decode": false})
 									r1_2 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true})  # cached
@@ -847,7 +741,7 @@ func TestHTTPSendCaching(t *testing.T) {
 		},
 		{
 			note: "http.send GET cache miss different headers",
-			ruleTemplate: `p = x { 
+			ruleTemplate: `p = x {
 									r1 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true, "headers": {"h1": "v1", "h2": "v2"}})
 									r2 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true, "headers": {"h2": "v2"}})
 									r2 = http.send({"method": "get", "url": "%URL%", "force_json_decode": true, "headers": {"h2": "v3"}})
@@ -861,7 +755,7 @@ func TestHTTPSendCaching(t *testing.T) {
 		},
 		{
 			note: "http.send POST cache miss different body",
-			ruleTemplate: `p = x { 
+			ruleTemplate: `p = x {
 									r1 = http.send({"method": "post", "url": "%URL%", "force_json_decode": true, "headers": {"h2": "v2"}, "body": "{\"foo\": 42}"})
 									r2 = http.send({"method": "post", "url": "%URL%", "force_json_decode": true, "headers": {"h2": "v3"}, "body": "{\"foo\": 23}"})
 									r1_2 = http.send({"method": "post", "url": "%URL%", "force_json_decode": true, "headers": {"h2": "v2"}, "body": "{\"foo\": 42}"})  # cached
